# Conflicts:
#	backend-nest/src/session/controller/session.controller.ts
#	frontend-vue/src/components/question/AnswerGroup.vue
#	frontend-vue/src/components/questionary/CreateQuestionary.vue
#	frontend-vue/src/components/questionary/QuestionnaryEdit.vue

+++ backend-nest/src/question/service/question.service.ts
@@ -42,48 +42,48 @@ export class QuestionService {
     return question;
   }
 
-  async deleteQuestions(questionnary: Questionnary) {
+  async deleteQuestions(questionnary : Questionnary) {
     const questions = await this.questionRepository.find({
       where: { questionnary },
     });
-    for (const question of questions) {
-      await this.answerRepository.delete({ question });
+    if (!!questions) {
+      for (const question of questions) {
+        await this.answerRepository.delete({ question });
+      }
+      await this.questionRepository.delete({ questionnary });
     }
-    await this.questionRepository.delete({ questionnary });
+    return !!questions;
   }
 
-  async findQuestion(questionnary: Questionnary) {
-    const questionsDB = await this.questionRepository.find({
-      where: { questionnary },
+  async findQuestion(idQuestion: number) {
+    return await this.questionRepository.findOne({
+      where: { id: idQuestion },
       relations: ['answers'],
     });
-
-    const questions: Question[] = [];
-    for (const q of questionsDB) {
-      const answers = q.answers.map((answerEnt) => {
-        const answer = new Answer();
-        answer.id = answerEnt.id;
-        answer.content = answerEnt.content;
-        answer.isCorrect = answerEnt.isCorrect;
-        answer.question = q;
-        return answer;
-      });
-
-      const question = {
-        id: q.id,
-        type: q.type,
-        content: q.content,
-        answers: answers,
-        questionnary: questionnary,
-      };
-
-      questions.push(question);
-    }
-
-    return questions;
   }
 
-  async deleteQuestion(questionnary: Questionnary, idQuestion: number) {
+  async findQuestions(questionnary: Questionnary) {
+    return await this.questionRepository.find({
+      where: { questionnary },
+    });
+  }
+
+  async findAnswers(idQuestion: number) {
+    const questionsDB = await this.questionRepository.findOne({
+      where: { id: idQuestion },
+      relations: ['answers'],
+    });
+    return questionsDB.answers.map((answerEnt) => {
+      const answer = new Answer();
+      answer.id = answerEnt.id;
+      answer.content = answerEnt.content;
+      answer.isCorrect = answerEnt.isCorrect;
+      answer.question = null;
+      return answer;
+    });
+  }
+
+  async deleteQuestion(questionnary : Questionnary, idQuestion: number) {
     const question = await this.questionRepository.findOne({
       where: { questionnary, id: idQuestion },
     });
@@ -96,7 +96,7 @@ export class QuestionService {
 
   async modifyQuestion(
     question: Question,
-    questionnary: Questionnary,
+    questionnary : Questionnary,
     idQuestion: number,
   ) {
     const questionDB = await this.questionRepository.findOne({
