+++ frontend-vue/src/components/questionary/QuestionnaryItem.vue
@@ -7,15 +7,32 @@
     <v-btn icon="edit" @click="modifyQuestionnary"></v-btn>
     <v-btn icon="delete" @click="deleteQuestionnary"></v-btn>
   </div>
+
+  <v-dialog v-model="alertQuestionnaryDelete" max-width="600">
+    <v-card>
+      <v-card-title class="headline">Confirmation</v-card-title>
+      <v-card-text>
+        Are you sure to delete the questionnary :
+        <b>" {{ questionnaryName.title }} "</b>
+        ?
+      </v-card-text>
+      <v-card-actions>
+        <v-btn @click="alertQuestionnaryDelete = false">Cancel</v-btn>
+        <v-btn @click="deleteQuestionnary">confirm</v-btn>
+      </v-card-actions>
+    </v-card>
+  </v-dialog>
 </template>
 
 <script>
   import { useQuestionnaryStore } from '@/stores/questionnaryStore';
 
   export default {
+    emits: ['nextQuestionE'],
     data() {
       return {
         iflist: true,
+        alertQuestionnaryDelete: false,
       };
     },
     setup() {
@@ -27,14 +44,20 @@
     props: {
       questionnaryId: { type: Number, default: null },
       questionnaryName: {},
+      author: {},
     },
     methods: {
       deleteQuestionnary() {
-        this.useQ.deleteQuestionnary(this.questionnaryId);
+        if (!this.alertQuestionnaryDelete) {
+          this.alertQuestionnaryDelete = true;
+        } else {
+          this.useQ.deleteQuestionnary(this.questionnaryId);
+        }
       },
       modifyQuestionnary() {
         this.useQ.idQuestionnary = this.questionnaryName.id;
         this.useQ.getQuestionnary();
+        this.useQ.getQuestions();
         this.$emit('nextQuestionE');
       },
       startQuestionnary() {
